requirements.txt
# Основні залежності для Telegram бота
aiogram>=3.4.0
anthropic>=0.25.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Обробка зображень
opencv-python>=4.8.0
Pillow>=10.0.0

# Базові математичні бібліотеки
numpy>=1.24.0
aiohttp>=3.8.0
